+++ schema/index.d.ts
@@ -1,0 +1,6 @@
+import {
+	Row,
+	Schema
+} from '../types.d';
+
+export default function convert(data: Row[], schema: Schema): object[];